// src/app/api/model/review.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AuthorDetails } from './authorDetails';


export interface Review { 
    author?: string;
    author_details?: AuthorDetails;
    content?: string;
    created_at?: string;
    id?: string;
    updated_at?: string;
    url?: string;
}