// src/app/api/model/reviewPage.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Review } from './review';


export interface ReviewPage { 
    id?: number;
    page?: number;
    results?: Array<Review>;
    total_pages?: number;
    total_results?: number;
}